cleanup(client): tidy up tx logic a bit
