+++ packages/client/src/tx/index.ts
@@ -1,0 +1,8 @@
+export type {
+  TxBroadcastEvent,
+  TxEvent,
+  TxEventsPayload,
+  Transaction,
+  TxEntry,
+} from "./types"
+export { createTxEntry, submit, submit$ } from "./tx"
